Refactor Obj-C generator
1. Parser is renamed to builder. We do not parse. We build structure.
2. ObjCProjectHandler renamed to ObjCEntitiesGenerator to eplicitely
specify what is the purpose of the handler.
3. Generator options extend base options class and allow to distinguish
what types should be processed.
4. Variuos codestyle corrections.
