+++ gradle-plugin/src/main/groovy/com/stanfy/helium/gradle/internal/SourceCodeGenerators.groovy
@@ -3,13 +3,13 @@ package com.stanfy.helium.gradle.internal
 import com.stanfy.helium.gradle.tasks.GenerateJavaConstantsTask
 import com.stanfy.helium.gradle.tasks.GenerateJavaEntitiesTask
 import com.stanfy.helium.gradle.tasks.GenerateJsonSchemaTask
-import com.stanfy.helium.gradle.tasks.GenerateObjcTask
+import com.stanfy.helium.gradle.tasks.GenerateObjcEntitiesTask
 import com.stanfy.helium.gradle.tasks.GenerateRetrofitTask
 import com.stanfy.helium.handler.codegen.java.constants.ConstantsGeneratorOptions
 import com.stanfy.helium.handler.codegen.java.entity.EntitiesGeneratorOptions
 import com.stanfy.helium.handler.codegen.java.retrofit.RetrofitGeneratorOptions
 import com.stanfy.helium.handler.codegen.json.schema.JsonSchemaGeneratorOptions
-import com.stanfy.helium.handler.codegen.objectivec.parser.options.DefaultObjCProjectParserOptions
+import com.stanfy.helium.handler.codegen.objectivec.ObjcEntitiesOptions
 
 /**
  * Description of source generators.
@@ -32,8 +32,8 @@ class SourceCodeGenerators {
           task: GenerateRetrofitTask
       ],
       objc: [
-          optionsFactory: { new DefaultObjCProjectParserOptions() },
-          task: GenerateObjcTask
+          optionsFactory: { new ObjcEntitiesOptions() },
+          task: GenerateObjcEntitiesTask
       ],
       jsonSchema: [
           optionsFactory: { JsonSchemaGeneratorOptions.defaultOptions() },
