+++ helium/src/main/groovy/com/stanfy/helium/handler/codegen/objectivec/mapper/ObjCMapper.java
@@ -1,7 +1,7 @@
 package com.stanfy.helium.handler.codegen.objectivec.mapper;
 
 import com.stanfy.helium.handler.codegen.objectivec.ObjCProject;
-import com.stanfy.helium.handler.codegen.objectivec.parser.options.ObjCProjectParserOptions;
+import com.stanfy.helium.handler.codegen.objectivec.ObjcEntitiesOptions;
 import com.stanfy.helium.model.Project;
 
 /**
@@ -15,5 +15,6 @@ public interface ObjCMapper {
   Updated specified |project| by adding additional file(s), those contain code that
   should map items from the DSL to the Classes, defined in project.
    */
-  void generateMappings(final ObjCProject project, final Project projectDSL, final ObjCProjectParserOptions options);
+  // TODO: replace with more generic options
+  void generateMappings(final ObjCProject project, final Project projectDSL, final ObjcEntitiesOptions options);
 }
