--- helium/src/main/groovy/com/stanfy/helium/handler/codegen/objectivec/parser/options/DefaultObjCProjectParserOptions.java
@@ -1,15 +1,0 @@
-package com.stanfy.helium.handler.codegen.objectivec.parser.options;
-
-public class DefaultObjCProjectParserOptions implements ObjCProjectParserOptions {
-
-  private String prefix = "HE";
-
-  public void setPrefix(final String prefix) {
-    this.prefix = prefix;
-  }
-
-  @Override
-  public String getPrefix() {
-    return prefix;
-  }
-}
