--- helium/src/main/groovy/com/stanfy/helium/handler/codegen/objectivec/parser/options/ObjCProjectParserOptions.java
@@ -1,16 +1,0 @@
-package com.stanfy.helium.handler.codegen.objectivec.parser.options;
-
-/**
- * Created by ptaykalo on 8/19/14.
- * Options those can be passed to the ObjCProjecteParser
- */
-public interface ObjCProjectParserOptions {
-
-  /*
-  Returns prefix for all files those would be generated
-   */
-  String getPrefix();
-
-}
-
-
