+++ helium/src/test/groovy/com/stanfy/helium/handler/codegen/objectivec/ObjCClassDefinitionSpec.groovy
@@ -9,36 +9,36 @@ import spock.lang.Specification
  */
 class ObjCClassDefinitionSpec extends Specification {
 
-    ObjCProject project;
-    ObjCHeaderFile headerFile
-    private String fileName
+  ObjCProject project;
+  ObjCHeaderFile headerFile
+  private String fileName
 
-    def setup() {
-        project = new ObjCProject();
-        fileName = "test"
-        headerFile = new ObjCHeaderFile(fileName);
+  def setup() {
+    project = new ObjCProject();
+    fileName = "test"
+    headerFile = new ObjCHeaderFile(fileName);
 
-    }
+  }
 
-    def "should add sourceParts"() {
-        when:
-        ObjCClassDefinition classDefinition = new ObjCClassDefinition(fileName);
-        ObjCPropertyDefinition propertyDefinition = new ObjCPropertyDefinition("name", "type");
-        classDefinition.addPropertyDefinition(propertyDefinition)
+  def "should add sourceParts"() {
+    when:
+    ObjCClassDefinition classDefinition = new ObjCClassDefinition(fileName);
+    ObjCPropertyDefinition propertyDefinition = new ObjCPropertyDefinition("name", "type");
+    classDefinition.addPropertyDefinition(propertyDefinition)
 
-        then:
-        classDefinition.getPropertyDefinitions().size() == 1
-    }
+    then:
+    classDefinition.getPropertyDefinitions().size() == 1
+  }
 
-    def "should generate contents of properties sourceParts when repersented as string"() {
-        when:
-        ObjCClassDefinition classDefinition = new ObjCClassDefinition(fileName);
-        ObjCPropertyDefinition propertyDefinition = new ObjCPropertyDefinition("name", "type");
-        classDefinition.addPropertyDefinition(propertyDefinition)
+  def "should generate contents of properties sourceParts when repersented as string"() {
+    when:
+    ObjCClassDefinition classDefinition = new ObjCClassDefinition(fileName);
+    ObjCPropertyDefinition propertyDefinition = new ObjCPropertyDefinition("name", "type");
+    classDefinition.addPropertyDefinition(propertyDefinition)
 
-        then:
-        classDefinition.asString().contains(propertyDefinition.asString());
-    }
+    then:
+    classDefinition.asString().contains(propertyDefinition.asString());
+  }
