+++ helium/src/test/groovy/com/stanfy/helium/handler/codegen/objectivec/generator/ObjCDefaultProjectGeneratorSpec.groovy
@@ -2,71 +2,69 @@ package com.stanfy.helium.handler.codegen.objectivec.generator
 
 import com.stanfy.helium.dsl.ProjectDsl
 import com.stanfy.helium.handler.codegen.objectivec.ObjCProjectGenerator
-import com.stanfy.helium.handler.codegen.objectivec.parser.DefaultObjCProjectParser
-import com.stanfy.helium.handler.codegen.objectivec.parser.options.DefaultObjCProjectParserOptions
+import com.stanfy.helium.handler.codegen.objectivec.ObjcEntitiesOptions
+import com.stanfy.helium.handler.codegen.objectivec.builder.DefaultObjCProjectBuilder
 import com.stanfy.helium.model.Type
 import org.apache.commons.io.FileUtils
-
 
 /**
  * Created by ptaykalo on 8/17/14.
  */
 class ObjCDefaultProjectGeneratorSpec extends ObjCProjectGeneratorSpec<ObjCProjectGenerator> {
 
-    ProjectDsl projectDSL;
-    DefaultObjCProjectParser parser;
+  ProjectDsl projectDSL;
+  DefaultObjCProjectBuilder parser;
 
-    def setup() {
+  def setup() {
 
-        projectDSL = new ProjectDsl()
+    projectDSL = new ProjectDsl()
 
-        projectDSL.typeResolver.registerNewType( new Type(name:"string"));
+    projectDSL.typeResolver.registerNewType( new Type(name:"string"));
 
-        when:
-        projectDSL.type "A" message {
-            name 'string'
-        };
-        projectDSL.type "B" message { }
-        projectDSL.type "C" message { }
+    when:
+    projectDSL.type "A" message {
+      name 'string'
+    };
+    projectDSL.type "B" message { }
+    projectDSL.type "C" message { }
 
-        // Parser
-        parser = new DefaultObjCProjectParser();
-        this.project = parser.parse(projectDSL);
+    // Parser
+    parser = new DefaultObjCProjectBuilder();
+    this.project = parser.build(projectDSL);
 
-        generator = new ObjCProjectGenerator(output, this.project);
-    }
+    generator = new ObjCProjectGenerator(output, this.project);
+  }
 
-    def "should generate implementation parts"() {
-        when:
-        generator.generate()
+  def "should generate implementation parts"() {
+    when:
+    generator.generate()
 
-        def fileAHeader = new File("$output/A.h")
-        def fileAHeaderContents = FileUtils.readFileToString(fileAHeader)
-        def fileAImpl = new File("$output/A.m")
-        def fileAImplContents = FileUtils.readFileToString(fileAImpl)
+    def fileAHeader = new File("$output/A.h")
+    def fileAHeaderContents = FileUtils.readFileToString(fileAHeader)
+    def fileAImpl = new File("$output/A.m")
+    def fileAImplContents = FileUtils.readFileToString(fileAImpl)
 
-        then:
-        fileAImplContents.contains("@implementation A")
-        fileAHeaderContents.contains("@interface A")
-    }
+    then:
+    fileAImplContents.contains("@implementation A")
+    fileAHeaderContents.contains("@interface A")
+  }
 
-    def "should generate implementation parts with prefixes from options"() {
-        when:
-        def options = new DefaultObjCProjectParserOptions()
-        this.project = parser.parse(projectDSL, options);
-        generator = new ObjCProjectGenerator(output, this.project);
+  def "should generate implementation parts with prefixes from options"() {
+    when:
+    def options = new ObjcEntitiesOptions()
+    this.project = parser.build(projectDSL, options);
+    generator = new ObjCProjectGenerator(output, this.project);
 
-        generator.generate();
+    generator.generate();
 
-        def fileAHeader = new File("$output/"+options.prefix +"A.h")
-        def fileAHeaderContents = FileUtils.readFileToString(fileAHeader)
-        def fileAImpl = new File("$output/"+options.prefix +"A.m")
-        def fileAImplContents = FileUtils.readFileToString(fileAImpl)
+    def fileAHeader = new File("$output/" + options.prefix +"A.h")
+    def fileAHeaderContents = FileUtils.readFileToString(fileAHeader)
+    def fileAImpl = new File("$output/" + options.prefix +"A.m")
+    def fileAImplContents = FileUtils.readFileToString(fileAImpl)
 
-        then:
-        fileAImplContents.contains("@implementation "+options.prefix +"A")
-        fileAHeaderContents.contains("@interface "+options.prefix +"A")
-    }
-
+    then:
+    fileAImplContents.contains("@implementation " + options.prefix +"A")
+    fileAHeaderContents.contains("@interface " + options.prefix +"A")
+  }
 
 }
