+++ helium/src/test/groovy/com/stanfy/helium/handler/codegen/objectivec/properties/ObjCProjectParserPropertiesSpec.groovy
@@ -2,112 +2,112 @@ package com.stanfy.helium.handler.codegen.objectivec.properties
 
 import com.stanfy.helium.dsl.ProjectDsl
 import com.stanfy.helium.handler.codegen.objectivec.ObjCProject
+import com.stanfy.helium.handler.codegen.objectivec.ObjcEntitiesOptions
+import com.stanfy.helium.handler.codegen.objectivec.builder.DefaultObjCProjectBuilder
 import com.stanfy.helium.handler.codegen.objectivec.file.ObjCClass
 import com.stanfy.helium.handler.codegen.objectivec.file.ObjCPropertyDefinition
-import com.stanfy.helium.handler.codegen.objectivec.parser.DefaultObjCProjectParser
-import com.stanfy.helium.handler.codegen.objectivec.parser.options.DefaultObjCProjectParserOptions
 import com.stanfy.helium.model.Type
 import spock.lang.Specification
 
 /**
  * Created by ptaykalo on 8/17/14.
  */
-class ObjCProjectParserPropertiesSpec extends Specification{
+class ObjCProjectParserPropertiesSpec extends Specification {
 
-    DefaultObjCProjectParser parser;
-    ProjectDsl project;
-    ObjCProject objCProject
-    DefaultObjCProjectParserOptions options
+  DefaultObjCProjectBuilder parser;
+  ProjectDsl project;
+  ObjCProject objCProject
+  ObjcEntitiesOptions options
 
-    def setup() {
-        project = new ProjectDsl()
-        parser = new DefaultObjCProjectParser()
-        options = new DefaultObjCProjectParserOptions();
-    }
+  def setup() {
+    project = new ProjectDsl()
+    parser = new DefaultObjCProjectBuilder()
+    options = new ObjcEntitiesOptions();
+  }
 
-    //        project.type "A" message { }
+  //        project.type "A" message { }
 
-    def "should generate ObjCProject with class"() {
-        given:
-        project.typeResolver.registerNewType( new Type(name:"string"));
+  def "should generate ObjCProject with class"() {
+    given:
+    project.typeResolver.registerNewType( new Type(name:"string"));
 
-        when:
-        project.type "A" message {
-            name 'string'
-        };
-        objCProject = parser.parse(project, options);
+    when:
+    project.type "A" message {
+      name 'string'
+    };
+    objCProject = parser.build(project, options);
 
-        then:
-        objCProject != null
-        objCProject.getClasses().size() == 1
-    }
+    then:
+    objCProject != null
+    objCProject.getClasses().size() == 1
+  }
 
-    def "should generate ObjCProject with class and property"() {
-        given:
-        project.typeResolver.registerNewType( new Type(name:"string"));
+  def "should generate ObjCProject with class and property"() {
+    given:
+    project.typeResolver.registerNewType( new Type(name:"string"));
 
-        when:
-        project.type "A" message {
-            name 'string'
-        };
-        objCProject = parser.parse(project, options);
-        ObjCClass aClass = objCProject.getClasses().get(0);
+    when:
+    project.type "A" message {
+      name 'string'
+    };
+    objCProject = parser.build(project, options);
+    ObjCClass aClass = objCProject.getClasses().get(0);
 
-        then:
-        aClass.definition != null
-        aClass.definition.propertyDefinitions.size() == 1
-    }
+    then:
+    aClass.definition != null
+    aClass.definition.propertyDefinitions.size() == 1
+  }
 
-    def "should generate properties with default nonatomic"() {
-        given:
-        project.typeResolver.registerNewType( new Type(name:"string"));
+  def "should generate properties with default nonatomic"() {
+    given:
+    project.typeResolver.registerNewType( new Type(name:"string"));
 
-        when:
-        project.type "A" message {
-            name 'string'
-        };
-        objCProject = parser.parse(project, options);
-        ObjCClass aClass = objCProject.getClasses().get(0);
+    when:
+    project.type "A" message {
+      name 'string'
+    };
+    objCProject = parser.build(project, options);
+    ObjCClass aClass = objCProject.getClasses().get(0);
 
-        then:
-        aClass.definition != null
-        aClass.definition.propertyDefinitions.size() == 1
-        aClass.definition.propertyDefinitions.get(0).getAtomicModifier() == ObjCPropertyDefinition.AtomicModifier.NONATOMIC;
-    }
+    then:
+    aClass.definition != null
+    aClass.definition.propertyDefinitions.size() == 1
+    aClass.definition.propertyDefinitions.get(0).getAtomicModifier() == ObjCPropertyDefinition.AtomicModifier.NONATOMIC;
+  }
 
-    def "should generate string properties with NSString type"() {
-        given:
-        project.typeResolver.registerNewType( new Type(name:"string"));
+  def "should generate string properties with NSString type"() {
+    given:
+    project.typeResolver.registerNewType( new Type(name:"string"));
 
-        when:
-        project.type "A" message {
-            name 'string'
-        };
-        objCProject = parser.parse(project, options);
-        ObjCClass aClass = objCProject.getClasses().get(0);
+    when:
+    project.type "A" message {
+      name 'string'
+    };
+    objCProject = parser.build(project, options);
+    ObjCClass aClass = objCProject.getClasses().get(0);
 
-        then:
-        aClass.definition != null
-        aClass.definition.propertyDefinitions.size() == 1
-        aClass.definition.propertyDefinitions.get(0).getType().equals("NSString *");
-    }
+    then:
+    aClass.definition != null
+    aClass.definition.propertyDefinitions.size() == 1
+    aClass.definition.propertyDefinitions.get(0).getType().equals("NSString *");
+  }
 
-    def "should generate string properties with NSString type and copy modifier"() {
-        given:
-        project.typeResolver.registerNewType( new Type(name:"string"));
+  def "should generate string properties with NSString type and copy modifier"() {
+    given:
+    project.typeResolver.registerNewType( new Type(name:"string"));
 
-        when:
-        project.type "A" message {
-            name 'string'
-        };
-        objCProject = parser.parse(project, options);
-        ObjCClass aClass = objCProject.getClasses().get(0);
+    when:
+    project.type "A" message {
+      name 'string'
+    };
+    objCProject = parser.build(project, options);
+    ObjCClass aClass = objCProject.getClasses().get(0);
 
-        then:
-        aClass.definition != null
-        aClass.definition.propertyDefinitions.size() == 1
-        // TODO - move to the options ?
-        aClass.definition.propertyDefinitions.get(0).getAccessModifier() == ObjCPropertyDefinition.AccessModifier.COPY;
-    }
+    then:
+    aClass.definition != null
+    aClass.definition.propertyDefinitions.size() == 1
+    // TODO - move to the options ?
+    aClass.definition.propertyDefinitions.get(0).getAccessModifier() == ObjCPropertyDefinition.AccessModifier.COPY;
+  }
 
 }
